refactor: move common logic from templatetag to utils file
