+++ solution/backend/regulations/utils/__init__.py
@@ -1,0 +1,1 @@
+from .link_statutes import *  # noqa
